Fix return data type for GetXmin() and GetXmax() methods
git-svn-id: https://svnsrv.desy.de/desy/h1fitter/trunk@334 b2a8d302-ca0a-427a-b7b9-081adbce06bc

+++ tools/include/PdfTable.h
@@ -35,9 +35,9 @@ class PdfTable  {
   virtual const double GetQ2(){return fQ2value;}
 
   // @brief Get Xmin value
-  const int GetXmin(){return fXmin;}
+  const double GetXmin(){return fXmin;}
   // @brief Get Xmax value
-  const int GetXmax(){return fXmax;}
+  const double GetXmax(){return fXmax;}
   
   // @brief Get PDF iPdf  at a given ix grid point 
   const double GetPDF(int iX, int iPdf);
